Move common code to common module
